chore: move ClassUtil into util package / adds a test for getInheritanceTree
